+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # Changlog
 
 ## v2.4.0
-* 
+* new darkMode property
+* new globalAlpha property
+* new setDarkMode() method
 
 ## v2.3.1
 * fixed concretejs pixel ratio issue